requirements.txt
anyio==3.6.2
black==23.3.0
certifi==2023.5.7
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Faker==18.7.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.1
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pydantic==1.10.7
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.22.0
